Implement sensible placement of display cards
Display cards (shown when hovering over a card) now appear above or below (based on the space available)  the card rather than always appearing in the middle of the screen.

Also disabled 'raycast target' on the card highlight part of card prefabs to prevent display cards flickering when hovering the edge of cards.