Improve performance of AI processing
Instead of serialising and deserialisng the action and the game state for every proposed action serialize every proposed action and the game state only once and then deserialise them for each proposed action. Unfortunately we have to deserialise every proposed action for each action game state pair but this is still a saving versus having to serialise the game state every time.